promptflow/src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
altgraph==0.17.3
amqp==5.1.1
anthropic==0.2.8
anyio==3.6.2
astroid==2.15.5
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
auto-py-to-exe==2.34.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.3.0
bleach==6.0.0
bottle==0.12.25
bottle-websocket==0.2.9
build==0.10.0
cachetools==5.3.0
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.26.3
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.1
customtkinter==5.1.2
darkdetect==0.8.0
decorator==5.1.1
dill==0.3.6
dnspython==2.3.0
docutils==0.19
Eel==0.16.0
elevenlabs==0.2.10
executing==1.2.0
fastapi==0.98.0
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
google-ai-generativelanguage==0.2.0
google-api-core==2.11.0
google-auth==2.18.1
google-generativeai==0.1.0rc3
google-search-results==2.4.2
googleapis-common-protos==1.59.0
googlesearch-python==1.2.3
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
iniconfig==2.0.0
InstructorEmbedding==1.0.0
ipython==8.13.2
isort==5.12.0
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==0.9.4
joblib==1.2.0
jsonschema==4.17.3
keyring==23.13.1
kombu==5.3.1
lazy-object-proxy==1.9.0
lit==16.0.2
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.5
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
myst-parser==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.6
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.2.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.1
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pylint==2.17.4
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pywin32-ctypes==0.2.0
PyYAML==6.0
questionary==1.5.1
readme-renderer==37.3
redis==4.6.0
redis-cli==0.1.2
regex==2023.5.5
requests==2.30.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sounddevice==0.4.6
soupsieve==2.4.1
Sphinx==6.2.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
starlette==0.27.0
svgwrite==1.4.3
sympy==1.11.1
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
tomlkit==0.11.8
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.0
triton==2.0.0
twine==4.0.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
whichcraft==0.6.1
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
